backend_sl/requirements.txt
asgiref==3.5.2
Django==4.0
sqlparse==0.4.2
tzdata==2022.2
django-autoslug==1.9.8
celery==5.2.7
Pillow==9.2.0
redis==4.3.4
django-cors-headers==3.13.0
httpie==3.2.1
psycopg2==2.9.3
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13
django-celery-beat==2.3.0
pytest-django==4.5.2
